requirements.txt
fireworks-python
pillow
python-dotenv
pydantic
streamlit